+++ MonsterGame_WithEngine/MonsterEngine/MonsterEngine-inl.h
@@ -2,7 +2,7 @@
 #pragma once
 
 inline int MonsterEngine::RandomIntInRange_Inclusive(int min, int max) {
-	return min + (rand() % (int)(max - min + 1));
+	return min + (rand() % (max - min + 1));
 }
