+++ MonsterGame_WithEngine/MonsterEngine/MonsterTransform.h
@@ -5,11 +5,11 @@
 class MonsterTransform
 {
 public:
-	MonsterTransform(double new_x, double new_y);
+	MonsterTransform(float new_x, float new_y);
 	~MonsterTransform();
 	inline const MonsterPoint2D& GetPosition() const;
 	void SetPosition(const MonsterPoint2D& new_pos);
-	void Translate(const double delta_x, const double delta_y);
+	void Translate(const float delta_x, const float delta_y);
 private:
 	MonsterPoint2D position_;
 };
